PR changes and undefined fix
